Enable triagebot no-merges check
This configuration will exclude rollup PRs and subtree/submodule sync
PRs from merge commit detection, and it will post the default warning
message and add the `has-merge-commits` and `S-waiting-on-author` labels
when merge commits are detected on other PRs.

The eventual vision is to have bors refuse to merge if the
`has-merge-commits` label is present. A reviewer can still
force the merge by removing that label if they so wish.